requirement.txt
flask==2.3.2
flask-cors==3.0.10
requests==2.31.0
mysql-connector-python==8.0.33
PyMuPDF
